Created custom Loader; kept previous version in case we want to use it again
